ServerSentEvents/absolute-url.html

<!DOCTYPE html>
<html>
<head>
	<title>The url attribute must return the absolute URL that resulted from resolving the value that was passed to the constructor:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The url attribute must return the absolute URL that resulted from resolving the value that was passed to the constructor:</h1>
	<div id="log"></div>
  <script>
  var url = "http://html5.mosquito-fp7.eu/messages/t0012";
    var source = new EventSource(url);
    test(function(){assert_equals(source.url, url)}, "The 'url' attribute is the absolute url.");
	test(function(){assert_equals(source.URL, url)}, "The 'URL' attribute is the absolute url.");

    source.addEventListener('message', function(e){
	}, false);
    source.addEventListener('open', function(e){ /* Connection opened. */
		
    }, false);
    source.addEventListener('error', function(e){
    				/* Connection closed. */
    		}, false);
  </script>
  </div>
</body>
</html>
